Comments to database and webpage
All the code for the comments. Can now comment on a post and have it appear however you will be redirected to /comment and won't see the comment until you refresh the page which will then force you to resubmit the form. To see your comment, just go to /home and it will appear.