Finished the issue #32
Can now go to the user page and see all of their posts. You can comment and follow from the page, we may want to add some html to make it more noticeable that it is that users page such as their name and profile picture at the top of the page right under the nav bar.